renamed files + placed in queries folder
